LayoutTests/media/video-src-invalid-poster-expected.txt

RUN(video.load())
EVENT(loadstart)
EXPECTED (video.clientWidth == '76') OK
EXPECTED (video.clientHeight == '103') OK
END OF TEST
